Extract disabled tool skill filtering
Move skill-to-tool resolution and disabled-tool checks into SkillDisabledToolFilter. This separates settings-based filtering from the remaining installation orchestration.
