+++ Packages/src/Editor/Infrastructure/SkillSetup/SkillDisabledToolFilter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 19cd7088f4c26496da42cfdbb611a4e7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
